requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
pydantic[email]
passlib[bcrypt]
bcrypt
python-multipart
python-jose[cryptography]

# npm create vite@latest frontend -- --template svelte
    # cd frontend
    # npm install
# npm install svelte-spa-router
# npm install bootstrap
# $ npm run dev
